+++ CHANGELOG.md
@@ -5,8 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Return the rules in their original order after applying the `@include` directive option `/optimiseDomainBlockingRules` [#217]
+
+[Unreleased]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.136...HEAD
 
 ## [v1.1.136] - 2024-05-22
+
 
 ### Changed